dialects: move FastMathFlag and FastMathAttrBase to xdsl.dialects.utils
